Option to get relative path to Textpattern installation.
Fixes theme handling. There are two ways to get the
relative path; resolving it by comparing two normalized
results, or using relative paths with DirectoryIterator.

If relative path is given to DirectoryIterator it should
give out a relative path, which we could then store.

We could actually update this all to work with relative
paths rather than absolute, but for now we are going
to go with resolving the paths where needed.

Fixes #1